[REF] iap, various: reorganize IAP module organization
PURPOSE

Clean and improve IAP services integration in Odoo. Clean IAP integrations
tools and improve data sharing with base modules, notably for CRM and Partner.

SPECIFICATIONS

In this commit we reorganize IAP module to better understand its content
and ease future cleaning

  * have models separated from tools;
  * rename some tools to find their grep. An iap_ prefix is added to ensure
    we don't clash with other global functions or methods;
  * perform some linting;

LINKS

Task ID 2248367
Community PR odoo/odoo#
Enterprise PR odoo/enterprise#
Upgrade PR odoo/upgrade#
IAP PR odoo/iap-apps#
