+++ addons/crm_iap_lead/models/crm_iap_lead_mining_request.py
@@ -4,8 +4,7 @@
 import logging
 
 from odoo import api, fields, models, _
-from odoo.exceptions import ValidationError
-from odoo.addons.iap import jsonrpc, InsufficientCreditError
+from odoo.addons.iap.tools import iap_tools
 
 _logger = logging.getLogger(__name__)
 
@@ -158,9 +157,9 @@ class CRMLeadMiningRequest(models.Model):
             'data': server_payload
         }
         try:
-            response = jsonrpc(endpoint, params=params, timeout=300)
+            response = iap_tools.iap_jsonrpc(endpoint, params=params, timeout=300)
             return response['data']
-        except InsufficientCreditError as e:
+        except iap_tools.InsufficientCreditError as e:
             self.error = 'Insufficient credits. Recharge your account and retry.'
             self.state = 'error'
             self._cr.commit()
