+++ addons/iap/models/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import iap
+from . import iap_account
 from . import res_config_settings
-
