+++ addons/snailmail/models/snailmail_letter.py
@@ -4,7 +4,7 @@ import re
 import base64
 
 from odoo import fields, models, api, _
-from odoo.addons.iap import jsonrpc
+from odoo.addons.iap.tools import iap_tools
 from odoo.tools.safe_eval import safe_eval
 
 DEFAULT_ENDPOINT = 'https://iap-snailmail.odoo.com'
@@ -333,7 +333,7 @@ class SnailmailLetter(models.Model):
         """
         endpoint = self.env['ir.config_parameter'].sudo().get_param('snailmail.endpoint', DEFAULT_ENDPOINT)
         params = self._snailmail_create('print')
-        response = jsonrpc(endpoint + PRINT_ENDPOINT, params=params)
+        response = iap_tools.iap_jsonrpc(endpoint + PRINT_ENDPOINT, params=params)
         for doc in response['request']['documents']:
             if doc.get('sent') and response['request_code'] == 200:
                 note = _('The document was correctly sent by post.<br>The tracking id is %s' % doc['send_id'])
